Add analyse pull request

+++ src/main/java/org/refactoringrefiner/api/Version.java
@@ -4,5 +4,5 @@ public interface Version {
 
     String getId();
 
-    int getTime();
+    long getTime();
 }
